+++ app/pytest.ini
@@ -1,21 +1,4 @@
 [pytest]
 DJANGO_SETTINGS_MODULE = DailyFavourite.settings
-python_files = tests.py test_*.py *_tests.py
-norecursedirs = .git .tox .venv env venv node_modules
-addopts = --cov=app --cov-report=html --cov-report=term-missing
-testpaths = app/tests
-filterwarnings =
-    ignore::DeprecationWarning
-    ignore::PendingDeprecationWarning
-    ignore::UserWarning
-    ignore::RuntimeWarning
-    ignore::ImportWarning
-    ignore::FutureWarning
-    ignore::UnicodeWarning
-    ignore::SyntaxWarning
-    ignore::ResourceWarning
-log_cli_level = INFO
-log_file_level = INFO
-log_file = app/tests/test.log
-log_format = %(asctime)s - %(name)s - %(levelname)s - %(message)s
-log_date_format = %Y-%m-%d %H:%M:%S
+python_files = tests.py test_*.py
+django_find_project = false
